spice-parent 15 --> spice-parent 20, while adjusting project core dependencies to match those present prior to this commit
The main reason to update parent is to pick up latest Sonatype confirmed working plugins related to release process.

This is also hopefully prepping this project for a last release in the 1.x line.

Subsequent releases is 2.x line should drop plexus-container-default with sisu-inject-plexus dependencies, and proceed with GAV adjustment to sisu-ldap-testsuite, to match half baked git repo name.